Fixes Instant/Timestamp conversion
An Instant may carry submillisecond data. So when it is added to Json, this information is reflected in the String conversion.
But the current regular expression in JDBCStatementHelper only allows for milliseconds precision.
Consequently, the value is not even identified as a timestamp, but as a String.

In addition to that, the current from/to SQL conversion methods simply erase submillisecond data, even though it's supported in the java.sql.Timestamp format.

SQLConvertTest has been updated with more test cases. Also it's been fixed because if something wasn't properly detected, it was simply returned as String and of course was equal to the original.