JDBCClient late initialization
Fixes #157

Currently the JDBCClient Datasource is created upfront (if not provided by the user).
As a consequence, if the provider does not lazily create JDBC connections it could:

- block the event loop, and/or
- throw runtime exceptions if the driver

With this patch, datasource creation is delayed until a connection is actually needed.
Besides, the creation process is executed on a worker thread.