Refactored JDBC action scheduling
# Removed runOnContext hops
This makes the code easier to reason about (in particular threading)

# Removed statement queues
These are no longer needed as we can now use the serialization provided by executeBlocking on the context bound to the JDBCConnectionImpl instance.

# Direct operation execution implementation change
Now it gets a connection then schedules execution on the worker pool.
Previously the operation was executed on the thread that acquires connections (thus serializing all operations on this JDBC client instance).
Also, this avoids to duplicate metrics and connection management (and tracing in the future).